+++ packages/zowe-explorer/__tests__/__unit__/extension.unit.test.ts
@@ -236,6 +236,7 @@ async function createGlobalMocks() {
             "zowe.jobs.ssoLogout",
             "zowe.jobs.startPolling",
             "zowe.jobs.stopPolling",
+            "zowe.jobs.cancelJob",
             "zowe.manualPoll",
             "zowe.updateSecureCredentials",
             "zowe.promptCredentials",