+++ packages/zowe-explorer/src/dataset/ZoweDatasetNode.ts
@@ -75,6 +75,9 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
         if (icon) {
             this.iconPath = icon.path;
         }
+        if (!globals.ISTHEIA && this.getParent() && contextually.isSession(this.getParent())) {
+            this.id = `${mParent?.id ?? mParent?.label?.toString() ?? "<root>"}.${this.label as string}`;
+        }
     }
 
     /**
@@ -98,7 +101,7 @@ export class ZoweDatasetNode extends ZoweTreeNode implements IZoweDatasetTreeNod
         if (!this.pattern && contextually.isSessionNotFav(this)) {
             return [
                 new ZoweDatasetNode(
-                    localize("getChildren.search", "Use the search button to display datasets"),
+                    localize("getChildren.search", "Use the search button to display data sets"),
                     vscode.TreeItemCollapsibleState.None,
                     this,
                     null,